fixed64: reject float operands to Fixed *,/,%,<<,>> and fix the finesine table
FRACUNIT is now the strong Fixed type, so `FRACUNIT * sin(x)` bound to
operator*(Fixed,int) and truncated sin -- which is in [0,1) across the first
quarter -- to 0. That zeroed the whole finesine/finecosine table and killed
movement, hitscan, monster sight and rendering at every non-cardinal angle.

Fix R_InitTables to spell the fill as a float multiply, double(FRACUNIT)*sin,
and close the whole class: delete `Fixed op <floating>` (both operand orders)
for *, /, %, << and >> as SFINAE templates that match only floating-point
types -- so the truncating form is a hard compile error while integer operands
still resolve to the scaling overloads with no long*Fixed ambiguity. Fixed-vs-
float comparisons and mixed Fixed/float ternaries are already rejected by the
explicit float ctor / explicit operator double.

Fix every flagged sibling site to preserve the original 32-bit math (int*float
promoted to a float multiply): wrap the Fixed operand in double(...) at each
call, and fix the three offending macros at their definitions -- PROP_FIXED_PARM,
GAMEINFOKEY_FIXED, AXERANGE, ANGLE_TO_FLOAT -- so one edit covers all invocations
instead of corrupting each expansion site.

Add regression tests: a compile-time contract (detection traits asserting
Fixed*double / /double / %double / <<double / <double / ternary are ill-formed
while integer forms stay valid) and a runtime guard that the finesine fill
produces the correct non-zero fixed values.

+++ src/zandronum/src/am_map.cpp
@@ -1949,8 +1949,8 @@ void AM_drawSubsectors()
 			{
 				AM_rotatePoint(&pt.x, &pt.y);
 			}
-			points[j].X = (float)(f_x + ((pt.x - m_x) * scale / float(1 << 24)));
-			points[j].Y = (float)(f_y + (f_h - (pt.y - m_y) * scale / float(1 << 24)));
+			points[j].X = (float)(f_x + (double((pt.x - m_x)) * double(scale) / float(1 << 24)));
+			points[j].Y = (float)(f_y + (f_h - double((pt.y - m_y)) * double(scale) / float(1 << 24)));
 		}
 		// For lighting and texture determination
 		sector_t *sec = Renderer->FakeFlat (subsectors[i].render_sector, &tempsec, &floorlight,	&ceilinglight, false);
@@ -2045,8 +2045,8 @@ void AM_drawSubsectors()
 			rotation += ANG90 - players[consoleplayer].camera->angle;
 			AM_rotatePoint(&originpt.x, &originpt.y);
 		}
-		originx = (double)(f_x + ((originpt.x - m_x) * scale / float(1 << 24)));
-		originy = (double)(f_y + (f_h - (originpt.y - m_y) * scale / float(1 << 24)));
+		originx = (double)(f_x + (double((originpt.x - m_x)) * double(scale) / float(1 << 24)));
+		originy = (double)(f_y + (f_h - double((originpt.y - m_y)) * double(scale) / float(1 << 24)));
 
 		// If this subsector has not actually been seen yet (because you are cheating
 		// to see it on the map), tint and desaturate it.
