+++ src/zandronum/src/g_shared/sbarinfo_commands.cpp
@@ -65,7 +65,7 @@ class CommandDrawImage : public SBarInfoCommandFlowControl
 			// We must calculate this per frame in order to prevent glitches with cl_capfps true.
 			fixed_t frameAlpha = block->Alpha();
 			if(alpha != FRACUNIT)
-				frameAlpha = fixed_t(((double) block->Alpha() / (double) FRACUNIT) * ((double) alpha / (double) OPAQUE) * FRACUNIT);
+				frameAlpha = fixed_t(double(((double) block->Alpha() / (double) FRACUNIT) * ((double) alpha / (double) OPAQUE)) * double(FRACUNIT));
 			
 			if(flags & DI_DRAWINBOX)
 			{
@@ -291,7 +291,7 @@ class CommandDrawImage : public SBarInfoCommandFlowControl
 					if (harmor->Slots[armorType] > 0 && harmor->SlotsIncrement[armorType] > 0)
 					{
 						//combine the alpha values
-						alpha = (int)(fixed_t(((double) alpha / (double) FRACUNIT) * ((double) MIN<fixed_t> (OPAQUE, Scale(harmor->Slots[armorType], OPAQUE, harmor->SlotsIncrement[armorType])) / (double) OPAQUE) * FRACUNIT));
+						alpha = (int)(fixed_t(double(((double) alpha / (double) FRACUNIT) * ((double) MIN<fixed_t> (OPAQUE, Scale(harmor->Slots[armorType], OPAQUE, harmor->SlotsIncrement[armorType])) / (double) OPAQUE)) * double(FRACUNIT)));
 						texture = statusBar->Images[image];
 					}
 					else
@@ -1694,7 +1694,7 @@ class CommandDrawSelectedInventory : public CommandDrawImage, private CommandDra
 				{
 					if(itemflash && itemflashFade)
 					{
-						fixed_t flashAlpha = fixed_t(((double) block->Alpha() / (double) FRACUNIT) * ((double) itemflashFade / (double) OPAQUE) * FRACUNIT);
+						fixed_t flashAlpha = fixed_t(double(((double) block->Alpha() / (double) FRACUNIT) * ((double) itemflashFade / (double) OPAQUE)) * double(FRACUNIT));
 						statusBar->DrawGraphic(statusBar->Images[statusBar->invBarOffset + imgCURSOR], imgx-4, imgy+2, block->XOffset(), block->YOffset(), (int)(flashAlpha), block->FullScreenOffsets(),
 							translatable, false, offset);
 					}
@@ -2211,7 +2211,7 @@ class CommandDrawInventoryBar : public SBarInfoCommand
 		
 			int bgalpha = block->Alpha();
 			if(translucent)
-				bgalpha = (int)(fixed_t((((double) block->Alpha() / (double) FRACUNIT) * ((double) HX_SHADOW / (double) FRACUNIT)) * FRACUNIT));
+				bgalpha = (int)(fixed_t(double((((double) block->Alpha() / (double) FRACUNIT) * ((double) HX_SHADOW / (double) FRACUNIT))) * double(FRACUNIT)));
 		
 			AInventory *item;
 			unsigned int i = 0;
@@ -3616,7 +3616,7 @@ class CommandAlpha : public SBarInfoMainBlock
 		void	Parse(FScanner &sc, bool fullScreenOffsets)
 		{
 			sc.MustGetToken(TK_FloatConst);
-			alpha = (int)(fixed_t(FRACUNIT * sc.Float));
+			alpha = (int)(fixed_t(double(FRACUNIT) * double(sc.Float)));
 
 			// We don't want to allow all the options of a regular main block
 			// so skip to the SBarInfoCommandFlowControl.
